Someday I'll get all the joins right...

+++ static/js/jamminc-browse.js
@@ -1,11 +1,15 @@
 
 $(function (event) {
     var delete_item = function (API_URL, id, name, $element) {
+        if (!confirm("Are you sure you want to delete this " + name + "?")) {
+            return;
+        }
+
         $.ajax({
             url: API_URL + "?" + $.param({ id: id }),
             type: "DELETE",
             success: function (data, textStatus, jqXHR) {
-                mw.flash(name + " deleted");
+                mw.flash(mw.capitalize(name) + " deleted");
                 $element.fadeOut();
             }
         });
@@ -16,10 +20,10 @@ $(function (event) {
             var type, name;
             if ($(this).hasClass("songs")) {
                 type = "songs";
-                name = "Song";
+                name = "song";
             } else if ($(this).hasClass("instruments")) {
                 type = "instruments";
-                name = "Instrument";
+                name = "instrument";
             }
             
             var API_URL = "/jamminc/music/" + type + ".json";
